move input validation routines into their own module
